-Changed float to double in class added csv file to main program folder

+++ workbook3/SearchInventoryMap/src/main/java/com/pluralsight/Product.java
@@ -4,9 +4,9 @@ public class Product {
 
     private int id;
     private String name;
-    private float price;
+    private double price;
 
-    public Product(int id, String name, float price) {
+    public Product(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -17,7 +17,7 @@ public class Product {
     public String getName() {
         return this.name;
     }
-    public float getPrice() {
+    public double getPrice() {
         return this.price;
     }
 }
